Forward original headers when processing includes
ESI originated requests now include the original cookies and
authorization headers. This allows us to use ESIs to include
user specific or restricted access information on pages.

Rather than forwarding all headers from the original request, this
filters out a whitelist of headers that are desirable to forward. For
example if while processing a POST request, the 'Content-Length' header
were to be forwarded on an ESI subrequest, it would cause the remote
server to wait for further content, and the request will timeout.